Rename data/event.js to data/event-deprecated.js
