+++ src/integrations/cms.js
@@ -5,7 +5,7 @@ import Config from "react-native-config";
 import { createClient } from "contentful/dist/contentful.browser.min";
 import { saveCmsData, loadCmsData } from "./storage";
 import type { Asset } from "../data/asset";
-import type { Event, FeaturedEvents } from "../data/event";
+import type { Event, FeaturedEvents } from "../data/event-deprecated";
 
 export type CmsEntry = Event | FeaturedEvents;
 export type CmsData = {
