+++ src/screens/EventsScreen/index.js
@@ -6,7 +6,7 @@ import type {
   EventCategoryName,
   SavedEvents,
   EventDays
-} from "../../data/event";
+} from "../../data/event-deprecated";
 import type { FieldRef } from "../../data/field-ref";
 import type { ImageSource } from "../../data/get-asset-source";
 import getAssetSource from "../../data/get-asset-source";
