+++ src/screens/FeaturedEventListScreen/component.js
@@ -2,7 +2,7 @@
 import React, { PureComponent } from "react";
 import { StyleSheet, View } from "react-native";
 import type { NavigationScreenProp } from "react-navigation";
-import type { SavedEvents, EventDays } from "../../data/event";
+import type { SavedEvents, EventDays } from "../../data/event-deprecated";
 import type { FieldRef } from "../../data/field-ref";
 import type { ImageSource } from "../../data/get-asset-source";
 import EventList from "../../components/EventList";
