+++ src/selectors/basic-event-filters.js
@@ -8,7 +8,7 @@ import {
 } from "../lib/date";
 import { selectEventIsFree } from "./event";
 import areaBoundaries from "../data/areas";
-import type { Event, EventCategoryName } from "../data/event";
+import type { Event, EventCategoryName } from "../data/event-deprecated";
 import type { DateRange, Time } from "../data/date-time";
 import type { Area, StringFilterSet } from "../data/event-filters";
