+++ src/selectors/basic-event-filters.test.js
@@ -7,7 +7,7 @@ import {
   buildStringSetFilter,
   buildAreaFilter
 } from "./basic-event-filters";
-import type { Event, EventCategoryName } from "../data/event";
+import type { Event, EventCategoryName } from "../data/event-deprecated";
 
 export type BuildEventArguments = {
   startTime?: string,
